map/requirements.txt
Pillow==5.0.0
python-dateutil==2.6.1
six==1.11.0
